修正

+++ app/Http/Controllers/PostController.php
@@ -6,6 +6,8 @@ use Illuminate\Http\Request;
 use App\Models\Post;
 use App\Http\Requests\PostRequest;
 use Illuminate\Support\Facades\Auth;
+use App\Models\Category;
+use App\Models\Like;
 
 class PostController extends Controller
 {
@@ -20,11 +22,6 @@ class PostController extends Controller
         // 投稿に関連するコメントをロード 
         $post->load('comments');
         return view('posts.show')->with(['post' => $post]);
-    }
-    
-    public function create()
-    {
-        return view('posts.create');
     }
     
     public function edit(Post $post)
@@ -51,8 +48,57 @@ class PostController extends Controller
     
     public function delete(Post $post)
     {
-        $post->delete();
+        // $post::where('user_id','Auth::id()');
+        if(Auth::id()==$post['user_id']){
+            $post->delete();
+        }
+        
         return redirect('/');
+    }
+    
+    // only()の引数内のメソッドはログイン時のみ有効
+      public function __construct()
+      {
+        $this->middleware(['auth', 'verified'])->only(['like', 'unlike']);
+      }
+    
+     /**
+      * 引数のIDに紐づくリプライにLIKEする
+      *
+      * @param $id リプライID
+      * @return \Illuminate\Http\RedirectResponse
+      */
+      public function like($id)
+      {
+        Like::create([
+          'post_id' => $id,
+          'user_id' => Auth::id(),
+        ]);
+    
+        session()->flash('success', 'You Liked the Post.');
+    
+        return redirect()->back();
+      }
+    
+      /**
+       * 引数のIDに紐づくリプライにUNLIKEする
+       *
+       * @param $id リプライID
+       * @return \Illuminate\Http\RedirectResponse
+       */
+      public function unlike($id)
+      {
+        $like = Like::where('post_id', $id)->where('user_id', Auth::id())->first();
+        $like->delete();
+    
+        session()->flash('success', 'You Unliked the Post.');
+    
+        return redirect()->back();
+      }
+      
+    public function category(Category $category)
+    {
+        return view('posts.create')->with(['categories' => $category->get()]);
     }
     
 }
